union_assemblies を別ファイルに切り出した
