+++ src/nasap_net/helpers/__init__.py
@@ -3,3 +3,4 @@ from .assembly_sort import sort_assemblies_by_component_kind_counts
 from .component_reindexing import reindex_components_in_assembly
 from .composition_formula import generate_composition_formula
 from .unique_id_validation import validate_unique_ids
+from .assembly_union import union_assemblies
